move tests to http.rs
